refactor: Controller uri 중복제거
