+++ src/main/java/techcourse/myblog/web/HomeController.java
@@ -1,0 +1,21 @@
+package techcourse.myblog.web;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import techcourse.myblog.domain.ArticleRepository;
+
+public class HomeController {
+
+    private final ArticleRepository articleRepository;
+
+    public HomeController(ArticleRepository articleRepository) {
+        this.articleRepository = articleRepository;
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        model.addAttribute("articles", articleRepository.findAll());
+        return "index";
+    }
+}
