+++ src/main/resources/templates/index.html
@@ -110,7 +110,7 @@
                 </div>
                 <div class="col-lg-3"></div>
             </div>
-            <form action="/writing">
+            <form action="/articles/writing">
                 <button class="btn btn-primary btn-rounded btn-icon fixed lh-2-4 right-50 bottom-50 width-50px height-50px">
                     <i class="ti-plus"></i>
                 </button>
